Tori/requirements.txt
# Core dependencies
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.0
duckdb==0.9.2
redis==5.0.1
numpy==1.26.0
scipy==1.11.3
scikit-learn==1.3.1
pykoopman==0.4.2
cvxpy==1.3.2
apscheduler==3.10.4
pydantic==2.5.0
orjson==3.9.10

# Metrics and monitoring
prometheus-client==0.17.1

# Visualization and web
fastapi==0.104.1
uvicorn==0.23.2
websockets==12.0
plotly==5.18.0
dash==2.14.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
tqdm==4.66.1
pyyaml==6.0.1
colorama==0.4.6
click==8.1.7
rich==13.6.0
humanize==4.8.0
uuid==1.30
shortuuid==1.0.11

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.10.1
isort==5.12.0
pylint==3.0.2
mypy==1.6.1
coverage==7.3.2
debugpy==1.8.0
watchdog==3.0.0
